sample: External memory FD
Actually two samples. One sample acts as exporter, exporting memory to
an opaque fd containing data about its color render target, while the
other sample acts as importer. The exporter waits for the importer to
connect via UNIX domain socket and then it send the opaque fd. The
importer connects to the exporter and retrieves the fd to use for
importing memory.